Update faker dependency
Hello,

Given the fact that the faker dependency is not available anymore since a vulnerability was introduced which led to the removal of the original git repository. 
I would like to ask if you could update the dependency to the new and stable one (you can check on this link https://www.npmjs.com/package/@faker-js/faker).
Thank you!